Split theme editor modal into focused components
theme-code-editor-modal.tsx had grown to 1,358 lines with two NiceModal
definitions, the recursive file-tree renderer, the top toolbar, the
full review-overlay modal, and a handful of helper types/constants all
inline. The util layer was already separated; the UI side now matches.

Extracted five sibling files plus a shared style constants module:

- theme-editor-confirm-modal.tsx: ThemeEditorConfirmModal NiceModal
- theme-editor-input-modal.tsx: ThemeEditorInputModal NiceModal
- theme-file-tree.tsx: ThemeFileTree component (also owns the TreeNode
  type, SelectedNode type, buildTree, sortTreeNodes, and the
  fileActionButtonClass style — all only used here)
- theme-editor-toolbar.tsx: ThemeEditorToolbar component (header bar
  with theme name, modified-count badge, Close, Save)
- theme-review-modal.tsx: ThemeReviewModal component for the "All
  changes" overlay, plus the ReviewItem type and the buildReviewItems
  / formatReviewSummary helpers. selectedReviewPath state and its
  auto-select effect moved inside since they are private to the review
  flow — the parent now only needs reviewItems, an onClose callback,
  onOpenInEditor, and onRevert.
- theme-editor-styles.ts: button-class strings shared between the
  toolbar and the review pane

Main file dropped from 1,358 to ~947 lines. Behaviour is unchanged;
the 27 existing theme acceptance tests, 169 admin-x-settings unit
tests, and tsc all pass without modification — these tests already
exercise every extracted surface (typing in the file tree, both
NiceModals via the save/discard/rename flows, Cmd+S and click Save on
the toolbar, opening the review overlay and reverting from it), so
the existing suite is the safety net for this refactor.
